Added option handling and tests for new crossref handling

+++ lib/dblp.rb
@@ -10,13 +10,18 @@ module Dblp
   class << self
 
     def run(file, options)
+
+      if !file || !File.exists?(file)
+        puts "You need to specify a filename"
+        exit
+      end
       
       # Clean file to parse
       file = File.basename(file, File.extname(file)) + ".aux"
       overall_size = 0
       
       parser = Dblp::Parser.new(file)
-      grabber = Dblp::Grabber.new
+      grabber = Dblp::Grabber.new(options)
       File.open(options.output, "w+") do |f|
         f.puts parser.parse.inject([]) {|m, l| 
             m << grabber.grab(l)
